requirements_fashion.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
matplotlib>=3.7.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
--find-links https://download.pytorch.org/whl/torch_stable.html 